project/requirements.txt
asgiref==3.6.0
Django==5.0 # take latests version or version 5
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2023.3
sqlparse==0.4.3
